replace -999 error coding with 0.0000999
eplace the -999 warning when there is a communication error with editing the number starting at the 4th decimal place and inserting 3 9s. Thus if there is an error for 2 or more reads in a row the output from the turbidimeter will be the last good reading with the number slightly modified. If the last good reading was 12.45638459, then the returned value will be 12.45699900. You can search for these in your data set to see if there were read errors, but they won't mess up PID control or graphs.